+++ Projects/UOContent/Items/Clothing/BaseClothing.Migrations.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace Server.Items;
 
 public partial class BaseClothing
@@ -100,5 +102,24 @@ public partial class BaseClothing
         }
 
         PlayerConstructed = GetSaveFlag(flags, OldSaveFlag.PlayerConstructed);
+    }
+
+    private static bool GetSaveFlag(OldSaveFlag flags, OldSaveFlag toGet) => (flags & toGet) != 0;
+
+    [Flags]
+    private enum OldSaveFlag
+    {
+        None = 0x00000000,
+        Resource = 0x00000001,
+        Attributes = 0x00000002,
+        ClothingAttributes = 0x00000004,
+        SkillBonuses = 0x00000008,
+        Resistances = 0x00000010,
+        MaxHitPoints = 0x00000020,
+        HitPoints = 0x00000040,
+        PlayerConstructed = 0x00000080,
+        Crafter = 0x00000100,
+        Quality = 0x00000200,
+        StrReq = 0x00000400
     }
 }
